Squashed commit of the following:
commit 357c0934be5bdef7faeca44a6d95c043e8e4e32d
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Wed Feb 26 19:13:55 2025 +0100

    Merge with main

commit e9723f8a82413ac99ddbb233876a12773bdea067
Merge: 83217cd 3b4e62a
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Wed Feb 26 19:03:31 2025 +0100

    Merge branch 'main' into ResourceSkeleton-THO-142

commit 3b4e62a767fd6b9c9b6b1ee5fa2b55d7cda318d1
Author: Marc Casanova <55861084+SirKaza@users.noreply.github.com>
Date:   Wed Feb 26 17:58:38 2025 +0100

    Bug/override THO-179 (#29)

    * all headers following CodingStyle, all files formated, some useless includes & files deleted or moved

    * camera & editor reload in project

    * camera & editor added filter

    * editorUI filter removed

    * added missing overrides

commit ee0e8decd50da73b2352f97d59797369c06e7eb3
Author: caspi99 <65339956+caspi99@users.noreply.github.com>
Date:   Mon Feb 24 23:54:20 2025 +0100

    Lights loading/saving THO-66 (#23)

    * Color Ambient and color intensity saving and loading

    * Added check to light loading

    * Load UID cubemapTexture

    * Texture loading

    * Some comments

    * Update LightsConfig.h

    * Added check nullptr and more encapsulation

    * Deleted unnecesary getters

commit 24bd79f7ad09f333b78bd519e60ff61bbd0a7550
Author: Marc López <88272783+MarcLopezS@users.noreply.github.com>
Date:   Mon Feb 24 22:34:50 2025 +0100

    Fixed Drag & Drop Scene GameObject (#24)

    Has been disabled drag and drop when the source gameObjects is the Scene GameObject (Root GameObject).

commit 83217cd5b3a1e484c3ebe6de6feee069e450ef95
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Sun Feb 23 17:38:18 2025 +0100

    Set local transform when instantiating the nodes

commit eba5b51ca0c9a475285d6d6a2461c373fe9da699
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Sat Feb 22 20:53:20 2025 +0100

    Save node transforms and draw gizmos

commit db2cbf78dfcdfec6e5c5b9c8448024af9f06d0ec
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Sat Feb 22 17:55:14 2025 +0100

    Render meshes on loading model

commit 4d9b522c14e3592751d05722768a0e8e1313f543
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Sat Feb 22 14:56:28 2025 +0100

    Save Model as JSON. Load model resource into Scene

commit e573f423838963fd778308b3aa9f6ad01f8f9723
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Thu Feb 20 21:42:35 2025 +0100

    Load Material from binary

commit 0f4720277468e81f7a656005baef9bd0f7f4dea5
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Thu Feb 20 17:39:24 2025 +0100

    [THO-142] Model Saved. WIP Loading Model Resource

commit afb669d6a65ce1a85d7efe6ce8debcb5dc2e4cc4
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Thu Feb 20 17:02:05 2025 +0100

    [THO-142] Load all .gltf nodes in pre-order

    Transform is not yet saved

commit 91f5cc0106ef856c2bef9d0f0d4158849cfaaf80
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Wed Feb 19 22:31:21 2025 +0100

    WIP Import model

commit 8c10f94075570d5308c2602e507191223588b6f5
Author: Marc Casanova <55861084+SirKaza@users.noreply.github.com>
Date:   Wed Feb 19 22:09:31 2025 +0100

    THO-167 Bug/stack windows (#22)

    * directionalLight when is nullptr fixed

    * bug fixed and menu bar polished

    * all build warnings fixed from first release

commit 263c9dcd57082ea37204619e65d3eb1b210b0b3d
Author: PolCurto <115176650+PolCurto@users.noreply.github.com>
Date:   Wed Feb 19 21:49:09 2025 +0100

    ResourceModel and ModelImporter created

+++ SobrassadaEngine/Utils/EngineTimer.cpp
@@ -3,13 +3,17 @@
 #include "Globals.h"
 #include "SDL.h"
 
-EngineTimer::EngineTimer() : time(0), startTime(0), deltaTime(0), isEnabled(false) {}
+EngineTimer::EngineTimer() : time(0), startTime(0), deltaTime(0), isEnabled(false)
+{
+}
 
-EngineTimer::~EngineTimer() {}
+EngineTimer::~EngineTimer()
+{
+}
 
 void EngineTimer::Start()
 {
-    startTime = SDL_GetTicks();
+    startTime = static_cast<float>(SDL_GetTicks());
     isEnabled = true;
 }
 
@@ -20,4 +24,7 @@ float EngineTimer::Tick()
     return deltaTime;
 }
 
-float EngineTimer::TicksSinceStartup() const { return SDL_GetTicks() - startTime; }
+float EngineTimer::TicksSinceStartup() const
+{
+    return SDL_GetTicks() - startTime;
+}
